requirements.txt
streamlit
tensorflow>=2.0
numpy
Pillow>=8.0.0
gdown
pathlib
torchvision
matplotlib
shape
keras
torch